moved around some stuff, nothing too big
it was pretty big lol
