+++ eucalyptus-core/src/lib.rs
@@ -1,14 +1,5 @@
-pub fn add(left: u64, right: u64) -> u64 {
-    left + right
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn it_works() {
-        let result = add(2, 2);
-        assert_eq!(result, 4);
-    }
-}
+pub mod logging;
+pub mod camera;
+pub mod states;
+pub mod utils;
+pub mod scripting;
