+++ eucalyptus-core/src/scripting/camera.rs
