+++ eucalyptus-core/src/scripting/entity.rs
