+++ eucalyptus-core/src/scripting/lighting.rs
