+++ eucalyptus-core/src/scripting/math.rs
