+++ eucalyptus-core/src/utils.rs
@@ -1,0 +1,48 @@
+use dropbear_engine::entity::Transform;
+use std::path::PathBuf;
+
+use crate::states::{ModelProperties, Node};
+
+pub const PROTO_TEXTURE: &[u8] = include_bytes!("../../resources/proto.png");
+
+pub fn search_nodes_recursively<'a, F>(nodes: &'a [Node], matcher: &F, results: &mut Vec<&'a Node>)
+where
+    F: Fn(&Node) -> bool,
+{
+    for node in nodes {
+        if matcher(node) {
+            results.push(node);
+        }
+        match node {
+            Node::File(_) => {}
+            Node::Folder(folder) => {
+                search_nodes_recursively(&folder.nodes, matcher, results);
+            }
+        }
+    }
+}
+
+/// Progress events for project creation
+pub enum ProjectProgress {
+    Step {
+        _progress: f32,
+        _message: String,
+    },
+    #[allow(dead_code)] // idk why its giving me this warning :(
+    Error(String),
+    Done,
+}
+
+pub enum ViewportMode {
+    None,
+    CameraMove,
+    Gizmo,
+}
+
+#[derive(Clone, Debug)]
+pub struct PendingSpawn {
+    pub asset_path: PathBuf,
+    pub asset_name: String,
+    pub transform: Transform,
+    pub properties: ModelProperties,
+}
