+++ eucalyptus-editor/src/editor/input.rs
@@ -9,6 +9,7 @@ use transform_gizmo_egui::GizmoMode;
 use winit::{
     dpi::PhysicalPosition, event::MouseButton, event_loop::ActiveEventLoop, keyboard::KeyCode,
 };
+use eucalyptus_core::success_without_console;
 
 impl Keyboard for Editor {
     fn key_down(&mut self, key: KeyCode, _event_loop: &ActiveEventLoop) {
@@ -32,8 +33,8 @@ impl Keyboard for Editor {
         match key {
             KeyCode::KeyG => {
                 if self.is_viewport_focused && !is_playing {
-                    self.viewport_mode = crate::utils::ViewportMode::Gizmo;
-                    crate::info!("Switched to Viewport::Gizmo");
+                    self.viewport_mode = ViewportMode::Gizmo;
+                    info!("Switched to Viewport::Gizmo");
 
                     if let Some(window) = &self.window {
                         window.set_cursor_visible(true);
@@ -44,8 +45,8 @@ impl Keyboard for Editor {
             }
             KeyCode::KeyF => {
                 if self.is_viewport_focused && !is_playing {
-                    self.viewport_mode = crate::utils::ViewportMode::CameraMove;
-                    crate::info!("Switched to Viewport::CameraMove");
+                    self.viewport_mode = ViewportMode::CameraMove;
+                    info!("Switched to Viewport::CameraMove");
                     if let Some(window) = &self.window {
                         window.set_cursor_visible(false);
 
@@ -65,7 +66,7 @@ impl Keyboard for Editor {
                     if let Some(_) = &self.selected_entity {
                         self.signal = Signal::Delete;
                     } else {
-                        crate::warn!("Failed to delete: No entity selected");
+                        warn!("Failed to delete: No entity selected");
                     }
                 } else {
                     self.input_state.pressed_keys.insert(key);
@@ -78,8 +79,8 @@ impl Keyboard for Editor {
                         log::debug!("Deselected entity");
                     }
                 } else if self.is_viewport_focused && !is_playing {
-                    self.viewport_mode = crate::utils::ViewportMode::None;
-                    crate::info!("Switched to Viewport::None");
+                    self.viewport_mode = ViewportMode::None;
+                    info!("Switched to Viewport::None");
                     if let Some(window) = &self.window {
                         window.set_cursor_visible(true);
                     }
@@ -92,14 +93,14 @@ impl Keyboard for Editor {
                     match self.save_project_config() {
                         Ok(_) => {}
                         Err(e) => {
-                            crate::fatal!("Error saving project: {}", e);
+                            fatal!("Error saving project: {}", e);
                         }
                     }
                     log::info!("Successfully saved project, about to quit...");
-                    crate::success_without_console!("Successfully saved project");
+                    success_without_console!("Successfully saved project");
                     self.scene_command = SceneCommand::Quit;
                 } else if is_playing {
-                    crate::warn!("Unable to save-quit project, please pause your playing state, then try again");
+                    warn!("Unable to save-quit project, please pause your playing state, then try again");
                 }
             }
             KeyCode::KeyC => {
@@ -120,22 +121,22 @@ impl Keyboard for Editor {
                                 };
                                 self.signal = Signal::Copy(s_entity);
 
-                                crate::info!("Copied!");
+                                info!("Copied!");
 
                                 log::debug!("Copied selected entity");
                             } else {
-                                crate::warn!(
+                                warn!(
                                     "Unable to copy entity: Unable to fetch world entity properties"
                                 );
                             }
                         } else {
-                            crate::warn!("Unable to copy entity: Unable to obtain lock");
+                            warn!("Unable to copy entity: Unable to obtain lock");
                         }
                     } else {
-                        crate::warn!("Unable to copy entity: None selected");
+                        warn!("Unable to copy entity: None selected");
                     }
                 } else if matches!(self.viewport_mode, ViewportMode::Gizmo) {
-                    crate::info!("GizmoMode set to scale");
+                    info!("GizmoMode set to scale");
                     self.gizmo_mode = GizmoMode::all_scale();
                 } else {
                     self.input_state.pressed_keys.insert(key);
@@ -148,7 +149,7 @@ impl Keyboard for Editor {
                             self.signal = Signal::Paste(entity.clone());
                         }
                         _ => {
-                            crate::warn!("Unable to paste: You haven't selected anything!");
+                            warn!("Unable to paste: You haven't selected anything!");
                         }
                     }
                 } 
@@ -161,14 +162,14 @@ impl Keyboard for Editor {
                     if !is_playing {
                         match self.save_project_config() {
                             Ok(_) => {
-                                crate::success!("Successfully saved project");
+                                success!("Successfully saved project");
                             }
                             Err(e) => {
-                                crate::fatal!("Error saving project: {}", e);
+                                fatal!("Error saving project: {}", e);
                             }
                         }
                     } else {
-                        crate::warn!("Unable to save project config, please quit your playing and try again");
+                        warn!("Unable to save project config, please quit your playing and try again");
                     }
                     
                 } else {
@@ -185,7 +186,7 @@ impl Keyboard for Editor {
                         self.signal = Signal::Undo;
                     }
                 } else if matches!(self.viewport_mode, ViewportMode::Gizmo) && !is_playing {
-                    crate::info!("GizmoMode set to translate");
+                    info!("GizmoMode set to translate");
                     self.gizmo_mode = GizmoMode::all_translate();
                 } else {
                     self.input_state.pressed_keys.insert(key);
@@ -202,7 +203,7 @@ impl Keyboard for Editor {
             }
             KeyCode::KeyX => {
                 if matches!(self.viewport_mode, ViewportMode::Gizmo) && !is_playing {
-                    crate::info!("GizmoMode set to rotate");
+                    info!("GizmoMode set to rotate");
                     self.gizmo_mode = GizmoMode::all_rotate();
                 } else {
                     self.input_state.pressed_keys.insert(key);
@@ -229,7 +230,7 @@ impl Keyboard for Editor {
 impl Mouse for Editor {
     fn mouse_move(&mut self, position: PhysicalPosition<f64>) {
         if (self.is_viewport_focused
-            && matches!(self.viewport_mode, crate::utils::ViewportMode::CameraMove))
+            && matches!(self.viewport_mode, ViewportMode::CameraMove))
             || (matches!(self.editor_state, EditorState::Playing) && !self.input_state.is_cursor_locked)
         {
             if let Some(window) = &self.window {
