--- eucalyptus/Cargo.toml
@@ -1,109 +1,0 @@
-[package]
-name = "eucalyptus"
-description = "A visual game editor for the dropbear game engine"
-default-run = "eucalyptus"
-exclude = ["resources"]
-
-version.workspace = true
-edition.workspace = true
-license-file.workspace = true
-repository.workspace = true
-readme.workspace = true
-
-[dependencies]
-anyhow = { workspace = true, optional = true }
-app_dirs2 = { workspace = true, optional = true }
-bincode = { workspace = true, optional = true }
-chrono = { workspace = true, optional = true }
-dropbear-engine = { workspace = true, optional = true }
-egui-toast-fork = { workspace = true, optional = true }
-egui = { workspace = true, optional = true }
-egui_dnd = { workspace = true, optional = true }
-egui_dock-fork = { workspace = true, optional = true }
-egui_extras = { workspace = true, optional = true }
-gilrs = { workspace = true, optional = true }
-git2 = { workspace = true, optional = true, features = ["vendored-openssl"]}
-glam = { workspace = true, optional = true }
-hecs = { workspace = true, optional = true }
-log = { workspace = true, optional = true }
-log-once = { workspace = true, optional = true }
-model_to_image = { workspace = true, optional = true }
-once_cell = { workspace = true, optional = true }
-open = { workspace = true, optional = true }
-parking_lot = { workspace = true, optional = true }
-ron = { workspace = true, optional = true }
-serde = { workspace = true, optional = true }
-tokio = { workspace = true, optional = true }
-transform-gizmo-egui = { workspace = true, optional = true }
-wgpu = { workspace = true, optional = true }
-winit = { workspace = true, optional = true }
-clap = { workspace = true, optional = true }
-walkdir = { workspace = true, optional = true }
-zip = { workspace = true, optional = true }
-bytemuck = "1.23.2"
-
-rustyscript.workspace = true
-
-[target.'cfg(not(target_os = "android"))'.dependencies]
-rfd = { workspace = true, optional = true }
-
-[features]
-editor = [
-    "anyhow",
-    "app_dirs2",
-    "bincode",
-    "chrono",
-    "dropbear-engine",
-    "egui-toast-fork",
-    "egui",
-    "egui_dnd",
-    "egui_dock-fork",
-    "egui_extras",
-    "gilrs",
-    "git2",
-    "glam",
-    "hecs",
-    "log",
-    "log-once",
-    "model_to_image",
-    "once_cell",
-    "open",
-    "parking_lot",
-    "rfd",
-    "ron",
-    "serde",
-    "tokio",
-    "transform-gizmo-egui",
-    "wgpu",
-    "winit",
-    "clap",
-    "walkdir",
-    "zip"
-]
-# default = ["editor"]
-
-data-only = [
-    "serde",
-    "bincode",
-    "anyhow",
-    "tokio",
-    "glam",
-    "log",
-    "dropbear-engine",
-    "winit",
-    "hecs",
-    "ron",
-    "log-once",
-    "once_cell",
-    "egui",
-    "chrono",
-    "parking_lot",
-    "walkdir",
-    "zip"
-]
-
-[build-dependencies]
-anyhow = "1.0"
-app_dirs2 = "2.5"
-reqwest = { version = "0.12", features = ["blocking"] }
-zip = "4.3"
