+++ resources/template.ts
